Reinstate old annotation behavior (without objDesc, but directly on prototype)
Initially, we had the mother class, DBusInterface have a field _ifaceDesc containing
the Maps for methods, properties and signals used for the introspection.
The problem was that when we were extending this class to create DBus interfaces, and
we used annotating functions like DBusMethod() to change the _ifaceDesc's Maps, it
was bubbling up to the Mother class and thus subsequent children classes had the same
_ifaceDesc.

To remedy to that problem, we annotated an external object, objDesc that was used and
copied by the constructor of the classes. But that was forcing the user to manually
create a objDesc.

Now we removed objDesc and we removed the _ifaceDesc from the DBusInterface mother
class. Instead, DBusMethod() (and alike) annotate the prototype of the children classes.
This way, we reinstated the old behavior.