Fix marshalling of variants
A variant of a struct must have signature starting with '(' -
multiple values in a single variant is invalid. Conversely, a
variant of a single value can have a signature with length > 1,
if the type is a composite type.
In turn, that means the whole expression made no sense and should be killed.